requirements.txt
PyYAML
tensorflow
multiprocess